OrderLogic

+++ app/src/main/java/ru/mirea/mobilefront/dto/OrderDto.java
@@ -6,6 +6,6 @@ import java.util.Map;
 
 @Data
 public class OrderDto {
-    private final Map<String, Double> bookList;
+    private final Map<String, Integer> bookList;
     private final double fullPrice;
 }
